case conversion operations: avoid mutating underlying string object
Because we cannot guarantee sole ownership of the underlying string object
attached to an event's field, mutating it is dangerous and can produce race
conditions when multiple operations in a pipeline mutate the same field in
different ways.

This patch choses instead to use non-destructive operations.

Using non-destructive operations incurs a cost overhead of a single ruby object
allocation and subsequent single ruby object dereference (which may or may not
make the original object eligible for Garbage Collection). This is very minimal
overhead compared to everything else that is happening in the pipeline and the
optimization is not worth the loss of safety.