Fix the merge function of the mutate filter
The dest_field_value was derived before the added fields were iterated upon. This caused the subsequent `event.set`'s, when iterating over multiple `added_field_value`'s, to set the `dest_field` based on the _initial_ `dest_field_value`. 

The logic should, instead, update the `dest_field_value` on each iteration of `added_field` so that it may continue to add to the `dest_field_value` instead of always only adding the last item in the list.